+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.1 under development
 
+- Consolidate Bootstrap5 views into `yiirocks/voyti-views-bootstrap5` package.
+
 ## 1.0.0 — August 20, 2026
 
 - Initial release.